Use Google Analytics. File way. File per profile (#298)
