+++ src/main/resources/application-dev.properties
@@ -14,3 +14,7 @@ app.site-title=Yalsee Dev
 
 # Short Domain (#288)
 app.shortDomain=${SHORT_DOMAIN:d.yls.ee}
+
+# SEO Stuff
+app.seo.ga.enabled=true
+app.seo.ga.file=gtag.dev.html
