+++ src/main/resources/application.properties
@@ -36,3 +36,7 @@ app.site-title=Yalsee Local Stand
 
 # Short Domain (#288) - l.yls.ee == 127.0.0.1
 app.shortDomain=${SHORT_DOMAIN:l.yls.ee:8080}
+
+# SEO Stuff
+app.seo.ga.enabled=false
+app.seo.ga.file=-
