+++ src/test/java/io/kyberorg/yalsee/test/ui/main/MainViewTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 
 import static com.codeborne.selenide.Condition.*;
 import static com.codeborne.selenide.Selenide.open;
-import static io.kyberorg.yalsee.test.pageobjects.MainViewPageObject.GOOGLE_ANALYTICS_CONTROL_SPAN;
 
 public class MainViewTest extends SelenideTest {
 
@@ -42,16 +41,5 @@ public class MainViewTest extends SelenideTest {
         String noColor = "rgba(0, 0, 0, 0)";
 
         Assertions.assertEquals(elementColor, noColor);
-    }
-
-    @Test
-    public void correctGoogleAnalyticsScriptLoadedAndHidden() {
-        GOOGLE_ANALYTICS_CONTROL_SPAN.should(exist);
-
-        GOOGLE_ANALYTICS_CONTROL_SPAN.shouldNotBe(visible);
-        GOOGLE_ANALYTICS_CONTROL_SPAN.shouldHave(attribute("aria-hidden","true"));
-
-        //TODO value has hardcoded html name, change it with TestEnv enum/class jne, which provides per env based info
-        GOOGLE_ANALYTICS_CONTROL_SPAN.shouldHave(attribute("aria-valuetext","gtag.dev.html"));
     }
 }
